+++ components/App.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react'
 import { View } from 'react-native'
 import { createSwitchNavigator, createStackNavigator } from 'react-navigation'
-import { AppProvider, AppConsumer } from './context'
+import { AppProvider, AppConsumer } from '../static/context'
 
-import AuthScreen from './components/screens/AuthScreen'
-import AppLoadingScreen from './components/screens/AppLoadingScreen'
-import AppScreen from './components/screens/AppScreen'
-import HeaderApp from './components/header/HeaderApp'
+import AuthScreen from './screens/AuthScreen'
+import AppLoadingScreen from './screens/AppLoadingScreen'
+import AppScreen from './screens/AppScreen'
+import HeaderApp from './header/HeaderApp'
 
-import styles from './static/styles'
+import styles from '../static/styles'
 
 const AppStack = createStackNavigator(
   { AppLoading: AppLoadingScreen, App: AppScreen },
